--- lib/features/main_navigation/widgets/nav_tab.dart
@@ -1,51 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:tiktok_clone/constants/gaps.dart';
-
-class NavTab extends StatelessWidget {
-  final String text;
-  final bool isSelected;
-  final IconData icon;
-  final Function onTap;
-
-  const NavTab({
-    Key? key,
-    required this.text,
-    required this.isSelected,
-    required this.icon,
-    required this.onTap,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      //Expanded+GestureDetector+Container(활성화된, 색을 지우면안됨)를 해줘야 선택영역이 넓어진다.
-      child: GestureDetector(
-        onTap: () => onTap(),
-        child: Container(
-          color: Colors.black,
-          child: AnimatedOpacity(
-            opacity: isSelected ? 1 : 0.6,
-            duration: Duration(milliseconds: 300),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                FaIcon(
-                  icon,
-                  color: Colors.white,
-                ),
-                Gaps.v5,
-                Text(
-                  text,
-                  style: TextStyle(
-                    color: Colors.white,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
